Revert "v0.7.0 - ack, speaker, package reorganization (#42)"
This reverts commit bae42340d27c9bacf313aaba06b695fa5032543f.
