+++ .github/workflows/pylint.yml
@@ -1,8 +1,6 @@
 name: Pylint
 
-on:
-  push:
-  pull_request:
+on: [push]
 
 jobs:
   build:
