+++ .pylintrc
@@ -553,8 +553,7 @@ function-naming-style=snake_case
 #function-rgx=
 
 # Good variable names which should always be accepted, separated by a comma.
-good-names=f,
-           i,
+good-names=i,
            j,
            k,
            x,
