+++ lib/wificom/hardware/nina_wifi.py
@@ -1,5 +1,5 @@
 '''
-wifi_nina.py
+wifi.py
 Handles the WiFi connnection for supported boards.
 Currently supported boards:
 	- Arduino Nano RP2040 Connect,
@@ -7,7 +7,7 @@ Currently supported boards:
 '''
 import busio
 from digitalio import DigitalInOut
-from wificom.import_secrets import secrets_wifi_ssid,secrets_wifi_password, \
+from wificom.common.import_secrets import secrets_wifi_ssid,secrets_wifi_password, \
 	secrets_mqtt_broker, \
 	secrets_mqtt_username, \
 	secrets_mqtt_password
@@ -60,6 +60,5 @@ class Wifi:
 			password=secrets_mqtt_password,
 		)
 
-		MQTT.set_socket(socket, self.esp)
-
-		return mqtt_client
+		# Return esp to use with mqtt client
+		return self.esp, mqtt_client
