+++ lib/wificom/hardware/nvm.py
